[script] refactoring: multi-visit
